# [Performance] Use new GCP custom images