<!-- Describe the changes in this PR -->
Use the new custom image for GCP's GPU VM creation. 

Benchmark results:
|  VM Type 💻   | Old Provision 🕐                 | New Provision 🕐     | % Improvement ✅ |
| --------------- | ---------------------------- | ----------- | ----------- |
| GCP GPU      | 3min 35s      | 1min 15s    | 65% |
| AWS GPU       | 3min 30s     | 2min 20s | 30% |
| GCP CPU       | 1min 45s   | 1min 10s | 32% |
* Provision Time = from start of launching an instance to instance setup and ready to run any user workloads (does not include time to search for the available instance)


<!-- Describe the tests ran -->
<!-- Unit tests (tests/test_*.py) are part of GitHub CI; below are tests that launch on the cloud. -->

Tested (run the relevant ones):

- [x] Code formatting: `bash format.sh`
- [x]  [GCP] Run all smoke tests `pytest tests/test_smoke.py --gcp`