@yika-luo 2024-10-04
This PR depends on another PR https://github.com/skypilot-org/skypilot-catalog/pull/86